Add THL to README.md
Looking at the millions of #machinelearning tweets already processed, noise is a little over 94%. Looking at the 8,000 daily tweets that are tagged #machineLearning, the site filters and ranks the most popular shared content in realtime. Machine learning’s zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, and will always be free to use.